fixes: >privateAliases add   | fixed (it was adding guild aliases beforehand, please check that someone didn't add their by accident) >reddit  | should no longer error for text-only posts >meme | should no longer error for text-only posts Custom command permissions now work again OsuCommand has an issue with floating bpm values breaking the command
new:
>anilist searchAnime <name*>     | gives max 10 results, respond with the number to get the result
>anilist searchManaga <name*>    | same as above
variable: {userMention} now supports user* arguments: {userMention:user*} (ex. {userMention:629490627469377536} or {userMention:ToxicMushroom})

+++ src/main/kotlin/me/melijn/melijnbot/internals/web/osu/OsuApi.kt
@@ -1,7 +1,7 @@
 package me.melijn.melijnbot.internals.web.osu
 
-import io.ktor.client.HttpClient
-import io.ktor.client.request.get
+import io.ktor.client.*
+import io.ktor.client.request.*
 import me.melijn.melijnbot.internals.translation.OSU_URL
 import net.dv8tion.jda.api.utils.data.DataArray
 
@@ -90,7 +90,7 @@ class OsuApi(val httpClient: HttpClient, private val apiKey: String) {
             beatMapJson.getString("artist"), // Creater of music/song
             beatMapJson.getString("beatmap_id").toLong(),
             beatMapJson.getString("beatmapset_id").toLong(),
-            beatMapJson.getString("bpm").toInt(),
+            beatMapJson.getString("bpm").toFloat().toInt(),
             beatMapJson.getString("creator"), // User who uploaded the map
             beatMapJson.getString("creator_id").toLong(),
             beatMapJson.getString("difficultyrating").toFloat(), // amount of stars
